py-srv/bin/requirements.txt
bottle
bottle-sqlalchemy
psycopg2-binary

